+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - #17: Bump rubycritic dependency to 2.4.1
 - Bump rspec dependency to 3.4.1
 - Bump activemodel dependency to 4.2.5
+- Reorganize how we run rubocop-rspec
 - #30: Allow to pass config file to brakeman
 - #18: Add bundler-audit which checks for vulnerable versions of gems
 - #21: Replace readme validator with required files validators, which allows us to